refactor: moved commander to core
